+++ src/config/options/titlesOptions/index.tsx
@@ -1,0 +1,16 @@
+const titlesOptions = [
+  {
+    id: "id",
+    titleName: "",
+    priority: 0,
+    value: "isActive",
+    type: "toggle",
+  },
+  {
+    id: "abbreviatedName",
+    titleName: "Rol",
+    priority: 1,
+    value: "abbreviatedName",
+  },
+];
+export { titlesOptions };
