+++ src/config/positions/table/index.tsx
@@ -1,9 +1,11 @@
 import { IPosition } from "@pages/positions/tabs/positionsTabs/outlets/addPosition/types";
 import { DetailsModal } from "@pages/positions/tabs/positionsTabs/forms/detailsModal";
-import { IAction } from "@pages/positions/tabs/positionsTabs/types";
+
 import { Edit } from "@pages/positions/tabs/positionsTabs/forms/edit";
-import { IEntry } from "@design/table/types";
+
 import { Delete } from "@pages/positions/tabs/positionsTabs/forms/delete";
+import { IEntry } from "@ptypes/table/IEntry";
+import { IAction } from "@ptypes/table/IAction";
 
 const titlesOptions = [
   {
