+++ src/design/modals/actionsModal/index.tsx
@@ -1,10 +1,11 @@
 import { MdClear } from "react-icons/md";
-import { IEntry } from "@design/table/types";
+import { IEntry } from "@ptypes/table/IEntry";
 import { Stack, Grid } from "@inubekit/inubekit";
 import { basic } from "@design/tokens";
-import { IActionsTable } from "@pages/positions/tabs/positionsTabs/types";
-import { IFormEntry } from "@design/templates/assignmentForm/types";
+import { IActionsTable } from "@ptypes/positions/table/IActionsTable";
+import { IFormEntry } from "@ptypes/assignmentForm/IFormEntry";
 import { StyledModal, StyledContentActions } from "./styles";
+
 interface IActionsModal {
   actions: IActionsTable[];
   entry: IFormEntry | IEntry;
