+++ src/hooks/design/useInitializerForm/index.ts
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { updateItemData } from "@mocks/utils/dataMock.service";
-import { IIUseInitializerForm } from "@design/forms/InitializerForm/types";
 import { IAssignmentFormEntry, IMessageState } from "@ptypes/positions/forms";
+import { IIUseInitializerForm } from "@ptypes/initializerForm/IIUseInitializerForm";
 
 const LoadingTimeout = 1500;
