+++ src/hooks/positions/useDeletePositions/index.ts
@@ -4,7 +4,7 @@ import { ISaveDataRequest } from "@ptypes/saveData/ISaveDataRequest";
 import { formatDate } from "@utils/date/formatDate";
 
 import { IAppData } from "@context/authAndDataProvider/types";
-import { IEntry } from "@design/table/types";
+import { IEntry } from "@ptypes/table/IEntry";
 
 const UseDeletePositions = (data: IEntry, appData: IAppData) => {
   const [showModal, setShowModal] = useState(false);
