+++ src/hooks/positions/useDeleteRequestInProgress/index.ts
@@ -1,4 +1,4 @@
-import { IEntry } from "@design/table/types";
+import { IEntry } from "@ptypes/table/IEntry";
 import { useState } from "react";
 
 const UseDeleteRequestInProgress = (
