+++ src/pages/positions/tabs/positionsTabs/outlets/addPosition/index.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { UseFetchRolesStaff } from "@hooks/positions/useFetchRolesStaff";
-import { IFormEntry } from "@design/templates/assignmentForm/types";
+
 import { UseAddStaffRoles } from "@hooks/positions/useAddStaffRoles";
 import { addStaffRolesSteps } from "@config/positions/addPositions/assisted";
 import { ISaveDataResponse } from "@ptypes/saveData/ISaveDataResponse";
@@ -10,6 +10,7 @@ import { AuthAndData } from "@context/authAndDataProvider";
 import { UseFetchAplicaionStaff } from "@hooks/positions/useAplication";
 import { IOptionInitialiceEntryApp } from "./types";
 import { AddStaffRolesUI } from "./interface";
+import { IFormEntry } from "@ptypes/assignmentForm/IFormEntry";
 const AddPosition = () => {
   const { rolesStaff } = UseFetchRolesStaff();
   const {
