+++ src/types/assignmentForm/IFormEntry/index.ts
@@ -1,0 +1,12 @@
+interface IFormEntry {
+  id: string;
+  value: string;
+  abbreviatedName?: string;
+  applicationName?: string;
+  isActive: boolean;
+  rolesStaff?: string;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  [key: string]: any;
+}
+
+export type { IFormEntry };
