+++ src/types/assignmentForm/IOptions/index.ts
@@ -1,0 +1,8 @@
+interface IOptions {
+  id: string;
+  label: string;
+  checked?: boolean;
+  onchange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+}
+
+export type { IOptions };
