+++ src/types/initializerForm/IInitializerFormUI/index.ts
@@ -1,5 +1,5 @@
-import { IFormEntry } from "@design/templates/assignmentForm/types";
 import { IOptionInitialiceEntryApp } from "@pages/positions/tabs/positionsTabs/outlets/addPosition/types";
+import { IFormEntry } from "@ptypes/assignmentForm/IFormEntry";
 import { IAssignmentFormEntry, IMessageState } from "@ptypes/positions/forms";
 
 interface IInitializerFormUI {
@@ -23,20 +23,4 @@ interface IInitializerFormUI {
   changeData?: IAssignmentFormEntry[];
 }
 
-interface IIUseInitializerForm {
-  dataOptionsForms: IAssignmentFormEntry[];
-  setSelectedToggle: React.Dispatch<
-    React.SetStateAction<IFormEntry[] | undefined>
-  >;
-  id?: string;
-  keyData?: string;
-  nameDB?: string;
-  property?: string;
-  propertyData?: string;
-  readOnly?: boolean;
-  withSubmitButtons?: boolean;
-  onHasChanges?: (hasChanges: boolean) => void;
-  dataOptionsValueSelect: IOptionInitialiceEntryApp[];
-}
-
-export type { IInitializerFormUI, IIUseInitializerForm };
+export type { IInitializerFormUI };
