+++ src/types/positions/table/IAction/index.ts
@@ -1,0 +1,9 @@
+import { IEntry } from "@ptypes/table/IEntry";
+
+interface IAction {
+  id: string;
+  actionName: string;
+  content: (entry: IEntry) => React.ReactNode;
+}
+
+export type { IAction };
