+++ src/types/positions/table/IActionsTable/index.ts
@@ -1,0 +1,10 @@
+import { IEntry } from "@ptypes/table/IEntry";
+
+interface IActionsTable {
+  id: string;
+  content: (entry: IEntry) => React.ReactNode;
+  mobilePriority?: boolean;
+  actionName?: string;
+  onClick?: (entry: IEntry) => void;
+}
+export type { IActionsTable };
