+++ src/types/positions/table/IPositions/index.ts
@@ -1,22 +1,7 @@
 import { PrivilegeOptionsConfig } from "@config/positions/tabs";
 import { ActionRenderer } from "@design/table/actionRenderer";
-import { IEntry } from "@design/table/types";
-
 import { IBusinessUnitsPortalStaffId } from "@ptypes/staffBusinessManagersId";
 
-interface IAction {
-  id: string;
-  actionName: string;
-  content: (entry: IEntry) => React.ReactNode;
-}
-
-interface IActionsTable {
-  id: string;
-  content: (entry: IEntry) => React.ReactNode;
-  mobilePriority?: boolean;
-  actionName?: string;
-  onClick?: (entry: IEntry) => void;
-}
 interface IPositions {
   handleSearchPositions: (e: React.ChangeEvent<HTMLInputElement>) => void;
   searchPosition: string;
@@ -39,4 +24,4 @@ interface IPositions {
   paginatedData: IBusinessUnitsPortalStaffId[];
 }
 
-export type { IAction, IPositions, IActionsTable };
+export type { IPositions };
