+++ src/types/positions/tabs/IPositionsUI/index.ts
@@ -1,0 +1,12 @@
+import { IEntry } from "@ptypes/table/IEntry";
+
+interface IPositionsUI {
+  isSelected: string;
+  handleTabChange: (id: string) => void;
+  smallScreenTab: boolean;
+  smallScreen: boolean;
+  data?: IEntry;
+  catalogName?: string;
+}
+
+export type { IPositionsUI };
