+++ src/types/positions/tabs/ITabConfig/index.ts
@@ -1,5 +1,3 @@
-import { IEntry } from "@design/table/types";
-
 interface ITabConfig {
   id: string;
   isDisabled: boolean;
@@ -7,13 +5,4 @@ interface ITabConfig {
   notificationIndicators?: number;
 }
 
-interface IPositionsUI {
-  isSelected: string;
-  handleTabChange: (id: string) => void;
-  smallScreenTab: boolean;
-  smallScreen: boolean;
-  data?: IEntry;
-  catalogName?: string;
-}
-
-export type { ITabConfig, IPositionsUI };
+export type { ITabConfig };
