+++ src/types/table/IBreakpoint/index.ts
@@ -1,0 +1,6 @@
+interface IBreakpoint {
+  breakpoint: string;
+  totalColumns: number;
+}
+
+export type { IBreakpoint };
