+++ src/types/table/IEntry/index.ts
@@ -1,0 +1,7 @@
+interface IEntry {
+  id?: string | number;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  [key: string]: any;
+}
+
+export type { IEntry };
