+++ src/types/table/ITitle/index.ts
@@ -1,0 +1,7 @@
+interface ITitle {
+  id: string;
+  titleName: string;
+  priority: number;
+}
+
+export type { ITitle };
