+++ src/types/users/usersTable/IUsersUI/index.ts
@@ -1,4 +1,4 @@
-import { IEntry } from "@design/table/types";
+import { IEntry } from "@ptypes/table/IEntry";
 
 interface IUsersUI {
   isSelected: string;
